refactorizo estructura
