--- src/cod/mvc/Main.java
@@ -1,7 +1,0 @@
-package cod.mvc;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
