+++ src/cod/mvc/model/Coche.java
@@ -1,4 +1,4 @@
-package cod.mvc;
+package cod.mvc.model;
 
 public class Coche {
     private String matricula;
