+++ test/src/test/java/TestModel.java
@@ -1,5 +1,5 @@
-import cod.mvc.Coche;
-import cod.mvc.Model;
+import cod.mvc.model.Coche;
+import cod.mvc.model.Model;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
